requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.10.6
pydantic-settings==2.0.3
python-dotenv==1.0.0
notion-client==2.0.0
google-generativeai==0.3.1
python-multipart==0.0.6
requests==2.31.0
jinja2==3.1.2
asyncio==3.4.3
gunicorn==21.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
google-cloud-storage==2.13.0
Markdown==3.5.1
